+++ bapi/rpc_test.go
