+++ rbot/bot_test.go
