+++ rbot/handler_test.go
