+++ rbot/pm_test.go
