+++ tests/premade-templates/go_module.txt
