+++ tests/premade-templates/js_module.txt
