+++ tests/premade-templates/py_module.txt
